all: Extract configuration structure to config.go
