+++ pkg/email/sendgrid/sendgrid.go
@@ -25,12 +25,6 @@ import (
 	"go.thethings.network/lorawan-stack/pkg/log"
 )
 
-// Config for the SendGrid email provider.
-type Config struct {
-	APIKey      string `name:"api-key" description:"The SendGrid API key to use"`
-	SandboxMode bool   `name:"sandbox" description:"Use SendGrid sandbox mode for testing"`
-}
-
 // SendGrid is the type that implements SendGrid as email provider.
 type SendGrid struct {
 	logger    log.Interface
