+++ pkg/web/oauthclient/oauthclient.go
@@ -26,20 +26,6 @@ import (
 	"golang.org/x/oauth2"
 )
 
-// Config is the configuration for the OAuth client.
-type Config struct {
-	AuthorizeURL string `name:"authorize-url" description:"The OAuth Authorize URL"`
-	LogoutURL    string `name:"logout-url" description:"The OAuth Logout URL"`
-	TokenURL     string `name:"token-url" description:"The OAuth Token Exchange URL"`
-	RootURL      string `name:"-"`
-
-	ClientID     string `name:"client-id" description:"The OAuth client ID"`
-	ClientSecret string `name:"client-secret" description:"The OAuth client secret" json:"-"`
-
-	StateCookieName string `name:"-"`
-	AuthCookieName  string `name:"-"`
-}
-
 // OAuthClient is the OAuth client component.
 type OAuthClient struct {
 	component *component.Component
